Projects/2.Learning-Assistant-Chatbot/requirements.txt
langchain
openai
faiss-cpu
sentence-transformers
tiktoken
python-dotenv
streamlit # For a simple web UI